댓글crud, 리뷰 디테일 조회 배포 v0.0.5

+++ src/main/java/com/book/heejinbook/dto/book/response/BookListResponse.java
@@ -13,6 +13,6 @@ public class BookListResponse {
     private String title;
     private String author;
     private String thumbnail;
-    private Integer reviewCount;
+    private Long reviewCount;
 
 }
